Remove contacts-admin frontend code
https://github.com/alphagov/contacts-admin/pull/249 moved responsibility for rendering the contacts index page from contacts-admin to finder-frontend. This commit removes all frontend-related code from contacts-admin so that it remains only as an admin app to add/edit/delete contacts. It also removes an old rake task for publishing the JSON for the index page to the Publishing API since there is now a new task at `lib/tasks/publish_finders.rake` which does so on behalf of finder-frontend.
