+++ Gemfile
@@ -5,7 +5,6 @@ gem "rails", "4.2.7.1"
 gem "active_hash", "1.4.1"
 gem "mysql2", "~> 0.4.4"
 gem "simple_form", "3.3.1"
-gem 'slimmer', "~> 9.5.0"
 gem "friendly_id", "5.1.0"
 gem "unicorn", "~> 5.1.0"
 gem "virtus", "~> 1.0.5"
