+++ README.md
@@ -1,44 +1,32 @@
 # Contacts app
 
-This app publishes contact information for a given
-organisation to GOV.UK.  It replaces the
-[HMRC contact us application](http://search2.hmrc.gov.uk/kb5/hmrc/contactus/home.page) and although it works for any organisation, only HMRC currently use it.
+This app publishes contact information for a given organisation to GOV.UK. It replaces the [HMRC contact us application](http://search2.hmrc.gov.uk/kb5/hmrc/contactus/home.page) and although it works for any organisation, only HMRC currently uses it.
 
 ## Live examples
 
-This application serves the index/filtering page for contacts, for example:
+[finder-frontend](https://github.com/alphagov/finder-frontend) displays the index page for an organisation's contacts:
 - https://www.gov.uk/government/organisations/hm-revenue-customs/contact
 
-Another application, [contacts-frontend](https://github.com/alphagov/contacts-frontend)
+[contacts-frontend](https://github.com/alphagov/contacts-frontend)
 displays the contacts themselves, fetching them from content store:
 - https://www.gov.uk/government/organisations/hm-revenue-customs/contact/alcohol-duties-national-registration-unit
 
 ## Technical documentation
 
-This is a Ruby on Rails application that serves 2 purposes:
+This is a Ruby on Rails application that provides a private admin UI for signon users with permission to allow them to manage the contacts for their organisation.
 
-1. It provides the public index of an organisation's contacts to users.
-2. It provides a private admin UI for signon users with permission to allow
-   them to manage the contacts for their organisation.
-
-The app has its own database (MySQL).  The frontend part of this app reads
-contacts directly from this database.  This is in contrast to the separate [contacts-frontend](https://github.com/alphagov/contacts-frontend) app which reads contacts from the [content-store](https://github.com/alphagov/content-store).  The admin UI part of this app handles publishing the contacts to the [publishing-api](https://github.com/alphagov/publishing-api) and [rummager](https://github.com/alphagov/rummager) so that they are present for [contacts-frontend](https://github.com/alphagov/contacts-frontend) to read.
+The app has its own database (MySQL), which is used by the admin UI. This is in contrast to the separate [contacts-frontend](https://github.com/alphagov/contacts-frontend) app which reads contacts from the [content-store](https://github.com/alphagov/content-store). The admin UI part of this app handles publishing the contacts to the [publishing-api](https://github.com/alphagov/publishing-api) and [rummager](https://github.com/alphagov/rummager) so that they are present for [finder-frontend](https://github.com/alphagov/finder-frontend) and [contacts-frontend](https://github.com/alphagov/contacts-frontend) to read.
 
 ### Dependencies
 
-* Ruby 2.1.4
+* Ruby 2.3.0
 * MySQL
-* [Rummager](https://github.com/alphagov/rummager) - to publish contacts so
-  they can be found by GOV.UK site search
-* [Whitehall](https://github.com/alphagov/whitehall)
-  1. to access the organisations API and maintain parity with the organisations
-     managed in whitehall.
-  2. to access the world locations API to provide the correct list of countries
-     for the country dropdown in the address form
-* [signon](https://github.com/alphagov/signonotron2) - for managing user auth in
-  the admin part
-* [static](https://github.com/alphagov/static) - for getting global GOV.UK
-  templates for the public part
+* [rummager](https://github.com/alphagov/rummager) - to publish contacts so
+  they can be found by GOV.UK site search and [finder-frontend](https://github.com/alphagov/finder-frontend)
+* [whitehall](https://github.com/alphagov/whitehall)
+  1. to access the organisations API and maintain parity with the organisations managed in whitehall.
+  2. to access the world locations API to provide the correct list of countries for the country dropdown in the address form
+* [signon](https://github.com/alphagov/signonotron2) - for managing user authentication
 * [content-store](https://github.com/alphagov/content-store) - for doing some
   command-line based tasks to withdraw and redirect a contact
 
@@ -46,12 +34,10 @@ contacts directly from this database.  This is in contrast to the separate [cont
 
 `./startup.sh`
 
-This runs `bundle install` to install dependencies and runs the app on port `3051`.  When using the GOV.UK development VM it will be available at http://contacts-admin.dev.gov.uk/.  The public frontend lives at `/contact/hm-revenue-customs/` and the admin UI at `/admin`
+This runs `bundle install` to install dependencies and runs the app on port `3051`. When using the GOV.UK development VM it will be available at http://contacts-admin.dev.gov.uk/admin.
 
-By default, this application uses the GOV.UK preview environment for assets. To
-run against a local version of static you need to set `STATIC_DEV` to
+By default, this application uses the GOV.UK preview environment for assets. To run against a local version of static, you need to set `STATIC_DEV` to
 "http://static.dev.gov.uk".
-
 
 ### Running the test suite
 
@@ -61,12 +47,11 @@ The tests in this project rely upon [govuk-content-schemas](http://github.com/al
 
 ### Tasks
 
-Rake tasks can be found in the standard location: `lib/tasks`.  The implementation of many of these can be found in `app/tasks` and `app/interactors` so they can also be run from the console.
+Rake tasks can be found in the standard location: `lib/tasks`. The implementation of many of these can be found in `app/tasks` and `app/interactors` so they can also be run from the console.
 
 ### Database setup
 
-The best way to get a database with good seed data is to use a dump from preview,
-alternatively you can load the database schema and use the old initial seed data:
+The best way to get a database with good seed data is to use a dump from preview; alternatively you can load the database schema and use the old initial seed data:
 
     ```
     bundle exec rake db:schema:load
