--- app/assets/javascripts/frontend.js
@@ -1,6 +1,0 @@
-//= require jquery_ujs
-//= require_tree ./frontend
-
-$(function(){
-  GOVUK.HmrcContacts.onLoad();
-});
