--- app/assets/stylesheets/frontend.css
@@ -1,4 +1,0 @@
-/*
- *= require_self
- *= require ./frontend/base
- */
