--- app/assets/stylesheets/frontend/base-ie6.scss
@@ -1,6 +1,0 @@
-// BASE STYLESHEET FOR IE 6 COMPILER
-
-$is-ie: true;
-$ie-version: 6;
-
-@import "base.scss";
