--- app/assets/stylesheets/frontend/base-ie7.scss
@@ -1,8 +1,0 @@
-// BASE STYLESHEET FOR IE 7 COMPILER
-
-$is-ie: true;
-$ie-version: 7;
-
-@import "base.scss";
-
-
