--- app/assets/stylesheets/frontend/base-ie8.scss
@@ -1,7 +1,0 @@
-// BASE STYLESHEET FOR IE 8 COMPILER
-
-$is-ie: true;
-$ie-version: 8;
-
-@import "base.scss";
-
