--- app/assets/stylesheets/frontend/base.scss
@@ -1,92 +1,0 @@
-// BASE STYLESHEET COMPILER
-
-
-
-// STYLEGUIDE
-// Mixins to be shared across gov.uk sites. Anything set in these
-// files can be used in the view files.
-
-// From the https://github.com/alphagov/govuk_frontend_toolkit
-@import "_colours.scss";
-@import "_conditionals.scss";
-@import "_typography.scss";
-@import "_shims.scss";
-@import "_css3.scss";
-@import "_device-pixels.scss";
-@import "design-patterns/_alpha-beta.scss";
-@import "measurements";
-
-@import "styleguide/_dimensions.scss";
-@import "styleguide/_font-stacks.scss";
-@import "styleguide/_typography.scss";
-@import "styleguide/_colours.scss";
-@import "styleguide/_button.scss";
-@import "styleguide/_float-fix.scss";
-@import "styleguide/_hidpi.scss";
-@import "styleguide/_right-to-left.scss";
-@import "styleguide/_govspeak.scss";
-
-/* conversion functions */
-
-  @function em($target-px, $current-px) {
-    @return $target-px * (1 / $current-px) * 1em;
-  }
-
-  @function px($value) {
-    @return $value * 1px;
-  }
-
-/* end conversion functions */
-
-// RESET
-// Flatten exising styles
-@import "reset/_govuk-overrides.scss";
-// overrides to the gov.uk header and footer
-@import "reset/_reset.scss";
-// reset styles applied to .hmrc-contacts-body & .hmrc-contacts-body *
-// so as to override global styles being set by gov.uk chrome. When
-// gov.uk styles are scoped, will want to change these selectors to
-// body and *
-
-// HELPERS
-// Styles for comment page elements
-.hmrc-contacts-body {
-  @import "helpers/_documents-index.scss";
-  @import "helpers/_documents-list.scss";
-  @import "helpers/_headings.scss";
-  @import "helpers/_heading-block.scss";
-  @import "helpers/_index-list.scss";
-  @import "helpers/_js-list-items.scss";
-  @import "helpers/_organisations.scss";
-  @import "helpers/_page-header.scss";
-}
-
-
-
-// VIEWS
-// Styles scoped to each template in /views
-
-.hmrc-contacts-body {
-  @import "views/_common.scss";
-  @import "views/_contacts.scss";
-  @import "views/_contact-show.scss";
-  @import "views/_documents-list.scss";
-  @import "views/_layouts.scss";
-  @import "views/_organisations.scss";
-}
-
-@-moz-document url-prefix(){
-  .single-row-select {
-    @media (min-width: 641px) {
-      text-indent: 0.01px !important;
-      text-overflow: '';
-      -moz-appearance: none;
-    }
-  }
-}
-
-
-
-// LAYOUTS
-// Styles scoped to a layout. Be careful.
-@import "layouts/_main.scss";
