--- app/assets/stylesheets/frontend/helpers/_heading-block.scss
@@ -1,65 +1,0 @@
-
-.heading-block {
-  margin-bottom: $gutter;
-  @extend %contain-floats;
-
-  h1 {
-    @include ig-core-27;
-    font-weight: bold;
-
-    @include media(tablet){
-      float: left;
-      clear: both;
-      width: 33.3%;
-      @include right-to-left {
-        float: right;
-      }
-    }
-  }
-  .content {
-    @include media(tablet){
-      float: right;
-      width: 66.6%;
-      @include right-to-left {
-        float: left;
-      }
-    }
-
-    article {
-      @include media(tablet){
-        padding-left: $gutter-half;
-
-        @include right-to-left {
-          padding-left: 0;
-          padding-right: $gutter-half;
-        }
-      }
-    }
-
-    .document-list {
-      @include media(tablet){
-        padding-left: $gutter-half;
-        margin-top: -$gutter-one-third;
-      }
-    }
-
-    .see-all {
-      @include core-19;
-      font-weight: bold;
-      padding-top: $gutter-one-third;
-
-      @include media(desktop) {
-        padding-left: $gutter-half;
-
-        @include right-to-left {
-          padding-left: $gutter-half;
-          padding-right: $gutter-half;
-        }
-      }
-      
-      a {
-        text-decoration: underline;
-      }
-    }
-  }
-}
