--- app/assets/stylesheets/frontend/helpers/_headings.scss
@@ -1,270 +1,0 @@
-.headings-block {
-  @extend %contain-floats;
-  margin-top: 0;
-
-  .inner-heading {
-    padding: 0 $gutter-half;
-  }
-
-  .heading {
-    padding: $gutter-half 0;
-
-    @include media(tablet){
-      padding: $gutter*1.5 0;
-    }
-
-    h1 {
-      @include core-36;
-      font-weight: bold;
-      padding-bottom: 20px;
-    }
-    .type {
-      @include core-27;
-      color: $secondary-text-colour;
-      margin-bottom: $gutter-one-sixth;
-    }
-  }
-  .heading-with-extra {
-    @extend .heading;
-
-    @include media(tablet){
-      float: left;
-      width: $two-thirds;
-      @include right-to-left {
-        float: right;
-      }
-    }
-  }
-  .heading-big {
-    @extend .heading;
-    h1 {
-      @include core-48;
-      font-weight: bold;
-    }
-  }
-  .heading-big-with-extra {
-    @extend .heading-with-extra;
-    @extend .heading-big;
-  }
-
-  .heading-extra {
-    margin: $gutter 0;
-    @include media(tablet) {
-      float: right;
-      width: $one-third;
-      margin-top: 0;
-      @include right-to-left {
-        float: left;
-      }
-    }
-
-    h4 {
-      @include ig-core-19;
-
-      @include media(tablet) {
-        position: absolute;
-        left: -9999em;
-      }
-    }
-
-    .all-services {
-      @include media(tablet) {
-        padding: 0 $gutter-two-thirds;
-      }
-    }
-
-    .top-tasks {
-      list-style: none;
-
-      @include media(tablet) {
-        padding-left: $gutter-two-thirds;
-      }
-
-      li {
-        @include core-16;
-        font-weight: bold;
-        padding: 3px 0;
-        @include media(tablet) {
-          padding: $gutter-one-sixth 0 0;
-        }
-
-        a[rel="external"] {
-          @include external-link-14-bold-no-hover;
-          @include media(tablet) {
-            @include external-link-16-bold-no-hover;
-          }
-        }
-      }
-    }
-
-    .available-languages {
-      margin: 0 0 $gutter-half 0;
-      ul {
-        padding: 0 0 $gutter-one-sixth 0;
-        li {
-          padding-right: $gutter-one-third;
-        }
-      }
-      @include media(tablet) {
-        margin: 0 0 $gutter-half $gutter-two-thirds;
-      }
-    }
-  }
-
-  .meta {
-    @extend %contain-floats;
-    display: block;
-    margin-bottom: $gutter;
-
-    @include media(tablet){
-      float: left;
-      width: $two-thirds;
-      @include right-to-left {
-        float: right;
-      }
-    }
-
-    dl {
-      display: inline;
-      @include ig-core-14;
-
-      dt {
-        @include bold;
-        float: left;
-        clear: left;
-        width: 40%;
-        @include media(tablet) {
-          width: auto;
-          min-width: ($gutter * 4);
-          padding-right: $gutter-one-third;
-
-          @include right-to-left {
-            padding-right: 0;
-            padding-left: $gutter-one-third;
-          }
-        }
-        @include right-to-left {
-          float: right;
-          clear: right;
-        }
-      }
-      dd {
-        float: left;
-        width: 60%;
-        @include media(desktop){
-          width: 70%;
-        }
-        &.document-worldwide-organisations {
-          width: 55%;
-        }
-        @include right-to-left {
-          float: right;
-        }
-        &::first-letter {
-          text-transform: uppercase;
-        }
-      }
-
-      a[rel="external"] {
-        @include external-link-12;
-        @include media(tablet) {
-          @include external-link-14;
-        }
-      }
-
-      .change-notes-title {
-        @include media(desktop){
-          clear: left;
-          @include right-to-left {
-            clear: right;
-          }
-        }
-      }
-    }
-  }
-  .available-languages {
-    ul {
-      @extend %contain-floats;
-      list-style: none;
-      padding: $gutter-half 0 $gutter-one-sixth;
-      border-bottom: 1px solid $border-colour;
-
-      li {
-        float: left;
-        display: block;
-        border-right: 1px solid $border-colour;
-        padding: 0 $gutter-one-third 0 0;
-        height: 16px;
-        margin: 3px $gutter-one-third 2px 0;
-        @include right-to-left {
-          float: right;
-        }
-
-        &.last {
-          border-right: 0;
-        }
-
-        a, span {
-          display: block;
-          margin-top: -1px;
-          @include core-16;
-        }
-        @include right-to-left{
-          border-left: 1px solid $border-colour;
-          border-right: 0;
-          margin: 3px 0 2px $gutter-one-third;
-          padding: 0 0 0 $gutter-one-third;
-          &.last {
-            border-left: 0;
-          }
-        }
-      }
-    }
-  }
-  .organisations-icon-list {
-    margin-top: $gutter*2;
-  }
-  .intro-paragraph {
-    padding: 0 $gutter-half;
-    margin-bottom: $gutter;
-    @include core-27;
-    max-width: 27em;
-
-    a {
-      text-decoration: underline;
-    }
-  }
-
-  .prominent-top-tasks {
-    @extend %contain-floats;
-    clear: both;
-    article {
-      @include media(tablet) {
-        float: left;
-        padding: $gutter 0;
-        width: $one-third;
-      }
-      .content {
-        padding: 0 0 $gutter-two-thirds 0;
-
-        h2 {
-          @include core-19;
-          font-weight: bold;
-        }
-        ul li {
-          font-weight: bold;
-          line-height: 1.25;
-          list-style: none;
-          padding: 0 0 $gutter-one-sixth 0;
-        }
-        @include media(tablet) {
-          padding: $gutter-half;
-        }
-      }
-    }
-    @include media(tablet) {
-      margin: 0 -$gutter-half;
-    }
-  }
-
-}
