--- app/assets/stylesheets/frontend/helpers/_index-list.scss
@@ -1,85 +1,0 @@
-.index-list-page {
-
-  .page-header {
-    padding: $gutter-half 0;
-
-    .title {
-      display: block;
-      @include core-36;
-      color: $grey-1;
-      padding-top: 5px;
-      padding-bottom: 2px;
-      margin-bottom: $gutter-one-sixth;
-    }
-
-    .label {
-      display: block;
-      @include core-36;
-      font-weight: bold;
-      input {
-        @include core-16;
-        position: relative;
-        padding: $gutter-one-sixth;
-        top: -$gutter-one-sixth;
-        @include media(tablet) {
-          top: -$gutter-one-third;
-        }
-      }
-      @include media(tablet) {
-        @include core-48;
-        font-weight: bold;
-      }
-    }
-
-    @include media(tablet){
-      padding: $gutter*1.5 0 ($gutter + $gutter-one-third);
-      width: 80%;
-    }
-
-  }
-
-  .js-filter-form {
-    display: none;
-
-    label {
-      display: inline-block;
-    }
-    input {
-      display: inline-block;
-      margin: 7px $gutter-one-third 3px;
-      width: 50%;
-      @include media(tablet) {
-        width: 30%;
-      }
-    }
-    div {
-      display: block;
-      input {
-        margin-left: 0;
-      }
-      @include media(tablet) {
-        display: inline;
-        input {
-          margin-left: $gutter-one-third;
-        }
-      }
-    }
-  }
-
-}
-
-.index-list-container {
-  margin-left: 33.333%;
-  
-  h3 {
-    @include ig-core-19;
-  }
-}
-
-.index-list {
-  li {
-    list-style: none;
-    border-bottom: 1px solid $border-colour;
-    @include ig-core-19;
-  }
-}
