--- app/assets/stylesheets/frontend/helpers/_js-list-items.scss
@@ -1,13 +1,0 @@
-.js-filter-no-results {
-  display: none;
-  width: 66%;
-  float: right;
-
-  &.reveal {
-    display: block;
-  }
-
-  p {
-    @include ig-core-19;
-  }
-}
