--- app/assets/stylesheets/frontend/helpers/_page-header.scss
@@ -1,51 +1,0 @@
-.page-header {
-  margin-top: 0;
-  background-repeat: no-repeat;
-  background-position: 0 8px;
-  margin-bottom: 10px;
-
-  h1 {
-    @include ig-core-48;
-    font-weight: bold;
-    .label {
-      display: block;
-    }
-  }
-
-  .explanation {
-    margin-top: $gutter-two-thirds*-1;
-    @include ig-core-19;
-  }
-
-  .ministers .minister {
-    list-style: none;
-  }
-
-  .extra-icons {
-    width: 100%;
-    overflow: hidden;
-    .organisation {
-      float: left;
-    }
-    .show-other-content {
-      float: left;
-      margin-left: $gutter-half;
-      @include ig-core-14;
-      padding: 3px 6px 0 4px;
-    }
-  }
-  .organisations-icon-list {
-    .hm-government {
-      padding-bottom: $gutter-one-sixth;
-      margin-bottom: $gutter-one-sixth;
-    }
-  }
-      
-}
-
-
-.show-other-content {
-  @include ig-core-14;
-  margin-left: 0.3em;
-  white-space: nowrap;
-}
