--- app/assets/stylesheets/frontend/layouts/_main.scss
@@ -1,56 +1,0 @@
-.hmrc-contacts-body {
-  #wrapper {
-    font-family: $NTA-Light;
-    position: relative;
-    overflow: visible;
-    margin: 0 auto;
-    padding-top: 0;
-    padding-bottom: $gutter;
-    background: transparent;
-    color: $text-colour;
-
-    @include media(desktop) {
-      padding-bottom: $gutter*2;
-    }
-  }
-
-  .hmrc-contacts-content {
-    position: relative;
-    margin: 0 auto;
-    width: auto;
-    max-width: $max-site-width + $gutter*2;
-
-    @extend %contain-floats;
-
-    @include ie-lte(6){
-      width: $max-site-width + $gutter*2;
-      padding-top: $gutter;
-    }
-  }
-
-  @include ie(6) {
-    #global-header .header-wrapper {
-      width: 990px;
-    }
-
-    #global-breadcrumb ol {
-      width: 930px;
-    }
-  }
-
-  .organisation-logo:link,
-  .organisation-logo:visited {
-    background: transparent;
-  }
-
-  #page {
-    @include right-to-left {
-      direction: rtl;
-    }
-  }
-
-}
-
-.js-enabled .js-hidden {
-  display: none !important;
-}
