--- app/assets/stylesheets/frontend/reset/_govuk-overrides.scss
@@ -1,19 +1,0 @@
-// override gov.uk page elements
-
-html {
-  background: white;
-}
-
-body,
-body.js-enabled {
-  padding: 0;
-}
-
-#global-header {
-  .header-proposition {
-    #proposition-menu {
-      margin-top: $gutter+$gutter-one-sixth+2px;
-      // 37px is the amount of margin required to line up the nav list with the top of the search box after removing Inside Gov heading
-    }
-  }
-}
