--- app/assets/stylesheets/frontend/styleguide/_dimensions.scss
@@ -1,9 +1,0 @@
-// DIMENSION UNITS
-// Whitespace values and maximum site width. These values can be
-// used in any file imported into this document.
-
-$max-site-width: 960px;
-
-$gutter-one-sixth:  $gutter*(1/6); // equivalent to 5px
-
-$three-quarters: ($full-width)-($one-quarter);
