--- app/assets/stylesheets/frontend/styleguide/_float-fix.scss
@@ -1,14 +1,0 @@
-// helpers for containing floats
-
-%contain-floats {
-  &:after {
-    content: ".";
-    display: block;
-    height: 0;
-    clear: both;
-    visibility: hidden;
-  }
-  @include ie-lte(7) {
-    zoom: 1;
-  }
-}
