--- app/assets/stylesheets/frontend/styleguide/_hidpi.scss
@@ -1,9 +1,0 @@
-@mixin hidpi {
-  @media (-webkit-min-device-pixel-ratio: 1.3),
-              (-o-min-device-pixel-ratio: 2.6/2),
-            (min--moz-device-pixel-ratio: 1.3),
-                 (min-device-pixel-ratio: 1.3),
-                         (min-resolution: 1.3dppx) {
-    @content;
-  }
-}
