--- app/assets/stylesheets/frontend/styleguide/_typography.scss
@@ -1,204 +1,0 @@
-//  GOV.UK typography palettes
-
-
-//  ANATOMY OF A TYPE STYLE
-//  -----------------------
-//  These are a collection of graphic styles. They are deliberately
-//  abstracted from semantic HTML context to enable flexible re-use.
-//  Although there is a lot of duplication within this file, as long
-//  as you GZIP your CSS it shouldnt cause any bloat.
-//
-//
-//  @mixin brand-45 {
-//  _ Style name refers to graphic style rather than semantic use
-//
-//      font-size: 45px;
-//
-//      line-height: (52/45);
-//      _ Unit-less, relative leading value.
-//        line-height (52px) / font-size (45px) = relative line-height
-//
-//      font-weight: 300;
-//      text-transform: none;
-//      _ Again, we cant assume anything about pre-exising CSS on the
-//        page, so setting these values (even if just resetting them
-//        to default values) will increase the robustness of the style.
-//
-//      padding-top: 7px;
-//      padding-bottom: 5px;
-//      _ Set top and bottom padding values to ensure an appropriate
-//        vertical measure before & after the text, ideally sitting it
-//        on the 10px baseline grid.
-//  }
-
-
-
-
-
-
-
-
-
-
-
-// CORE FONTS - NEW TRANSPORT
-
-// Display size used for infographics and pull-quotes/splashes
-@mixin ig-core-80($line-height: (80 / 80), $line-height-640: (55 / 53)) {
-  font-family: $NTA-Light;
-  @if $is-print == false {
-    font-size: 80px;
-  } @else {
-    font-size: 28pt;
-  }
-  line-height: $line-height;
-  font-weight: 400;
-  text-transform: none;
-  @media (max-width: 640px) {
-    font-size: 53px;
-    line-height: $line-height-640;
-  }
-}
-
-// Suitable for page headings
-@mixin ig-core-48 {
-  font-family: $NTA-Light;
-  font-size: 48px;
-  line-height: 1.0416666667; // 50 / 48
-  font-weight: 400;
-  text-transform: none;
-  padding-top: 12px;
-  padding-bottom: 8px;
-  @media (max-width: 640px) {
-    font-size: 25px;
-    line-height: (30/25);
-    padding-top: 8px;
-    padding-bottom: 7px;
-  }
-}
-
-// Suitable for page headings
-@mixin ig-core-36 {
-  font-family: $NTA-Light;
-  font-size: 36px;
-  line-height: 1.1111111111; // 40/36
-  font-weight: 400;
-  text-transform: none;
-  padding-top: 6px;
-  padding-bottom: 9px;
-  @media (max-width: 640px) {
-    font-size: 24px;
-    line-height: (25/24);
-    padding-top: 8px;
-    padding-bottom: 7px;
-  }
-}
-
-// Suitable for body headings
-@mixin ig-core-27 {
-  font-family: $NTA-Light;
-  font-size: 27px;
-  line-height: 1.1111111111; // 30px
-  font-weight: 400;
-  text-transform: none;
-  padding-top: 2px;
-  padding-bottom: 3px;
-  @media (max-width: 640px) {
-    font-size: 19px;
-    line-height: (25/19);
-    padding-top: 7px;
-    padding-bottom: 3px;
-  }
-}
-
-// Suitable for short body copy
-@mixin ig-core-24 {
-  font-family: $NTA-Light;
-  font-size: 24px;
-  line-height: (30/24); /* 30px */
-  font-weight: 400;
-  text-transform: none;
-  padding-top: 3px;
-  padding-bottom: 2px;
-  @media (max-width: 640px) {
-    font-size: 17px;
-    line-height: (25/17);
-    padding-top: 7px;
-    padding-bottom: 7px;
-  }
-}
-
-// Suitable for long body copy
-@mixin ig-core-19 {
-  font-family: $NTA-Light;
-  font-size: 19px;
-  line-height: (25/19); /* 25px */
-  font-weight: 400;
-  text-transform: none;
-  padding-top: 7px;
-  padding-bottom: 3px;
-  @media (max-width: 640px) {
-    font-size: 13px;
-    line-height: (20/13);
-    padding-top: 6px;
-    padding-bottom: 4px;
-  }
-}
-
-// Suitable body pullouts and asides
-@mixin ig-core-16 {
-  font-family: $NTA-Light;
-  font-size: 16px;
-  line-height: (20/16); /* 20px */
-  font-weight: 400;
-  text-transform: none;
-  padding-top: 5px;
-  padding-bottom: 5px;
-  @media (max-width: 640px) {
-    font-size: 13px;
-    line-height: (20/13);
-    padding-top: 6px;
-    padding-bottom: 4px;
-  }
-}
-
-// Suitable for captions, buttons etc
-@mixin ig-core-14 {
-  font-family: $NTA-Light;
-  font-size: 14px;
-  line-height: 1.4285714286; // 20 / 14
-  font-weight: 400;
-  text-transform: none;
-  padding-top: 6px;
-  padding-bottom: 4px;
-  @media (max-width: 640px) {
-    font-size: 12px;
-    line-height: (15/12);
-    padding-top: 4px;
-    padding-bottom: 1px;
-  }
-}
-
-@mixin bold {
-//  font-weight: 800;
-}
-
-@mixin light {
-  font-weight: 400;
-  color: $secondary-text-colour;
-}
-
-// TODO (@bfirsh): refactor frontend_toolkit external link mixin to work with 
-// white icons: https://github.com/alphagov/govuk_frontend_toolkit/pull/51
-@mixin external-link-white {
-  &:after {
-    background-image: image-url("external-link-white-12x12.png");
-    background-repeat: no-repeat;
-
-    @include device-pixel-ratio() {
-      background-image: image-url("external-link-white-24x24.png");
-      background-size: 12px 400px;
-    }
-  }
-}
-
