--- app/assets/stylesheets/frontend/views/_common.scss
@@ -1,127 +1,0 @@
-a {
-  text-decoration: none;
-
-  &:hover {
-    text-decoration: underline;
-  }
-
-  &.link-underline {
-    text-decoration: underline;
-
-    &:hover {
-      text-decoration: none;
-    }
-  }
-
-  &.button {
-    text-decoration: none;
-  }
-}
-
-ul {
-  padding-left: 15px;
-}
-
-input[type="text"],
-input[type="search"],
-input[type="number"],
-input[type="url"],
-input[type="password"],
-input[type="email"] {
-  -webkit-appearance: none;
-  @include border-radius(0);
-}
-
-.page-header h1 {
-  @media (max-width: 641px) {
-    font-size: em(32, 16);
-    line-height: 1.09375;
-  }
-}
-
-.hmrc-search {
-  .searchfield {
-    position: relative;
-    
-    .search-input {
-      position: static;
-      margin: 0;
-      padding-right: 50px;
-
-      @include ie(7) {
-        padding-right: 62px;
-      }
-
-      @include ie(6) {
-        border: 1px solid #BFC1C3;
-        border-right: none;
-        margin-right: $gutter-one-third;
-      }
-    }
-
-    .search-submit {
-      position: absolute;
-      bottom: 0;
-      right: 0;
-      height: 50px;
-      width: 50px;
-      overflow: visible;
-
-      @include ie(6) {
-        top: 0;
-        right: $gutter;
-      }
-    }
-
-    input {
-      @include core-19;
-      @include box-sizing(border-box);
-      width: 100%;
-      margin: 0;
-      padding: 0 0 0 10px;
-      display: block;
-      height: 50px;
-      border-color: $border-colour;
-      border-right: none;
-      line-height: 48px!important;
-      -webkit-appearance: none;
-      @include border-radius(0);
-
-      &::-webkit-search-cancel-button {
-        -webkit-appearance: none;
-      }
-
-      @include ie-lte(7) {
-        height: 48px;
-        margin-bottom: -1px;
-      }
-
-      @include ie(6) {
-        border: none;
-      }
-    }
-
-    button {
-      @include border-radius(0);
-      @include box-shadow(none);
-      position: absolute;
-      z-index: 4;
-      right: 0;
-      top: 0;
-      width: 50px;
-      height: 50px;
-      border: none;
-      overflow: hidden;
-      text-indent: -5000px;
-      color: $white;
-      background-color: $hm-revenue-customs;
-      background-image: image-url("search/search-button.png");
-      background-repeat: no-repeat;
-      background-position: -12.5% 50%;
-      
-      &:hover {
-        background-color: lighten($hm-revenue-customs, 5%);
-      }
-    }
-  }
-}
