--- app/assets/stylesheets/frontend/views/_contacts.scss
@@ -1,190 +1,0 @@
-.contacts-page {
-  .filter-block {
-    margin-bottom: $gutter;
-  }
-
-  .hmrc-search {
-    @media (min-width: 769px) {
-      margin-bottom: $gutter;
-    }
-
-    form {
-      display: block !important;
-    }
-  }
-
-  .organisations-index .page-header {
-    width: auto;
-    padding-top: $gutter-half;
-
-    .label input {
-      width: 50%;
-      margin: 7px 10px 3px;
-
-      @include media(mobile) {
-        width: auto;
-      }
-
-      @include ie-lte(7) {
-        position: relative;
-        top: 5px;
-        padding-top: 5px;
-        padding-bottom: 10px;
-      }
-    }
-  }
-
-  .heading-block {
-    clear: both;
-    margin-bottom: $gutter*2;
-
-    @include ie(6) {
-      .type-heading,
-      .content {
-        float: left;
-      }
-
-      .type-heading {
-        width: 33.3%;
-      }
-
-      .content {
-        width: 64.6%;
-      }
-    }
-
-    .type-heading h1 {
-      padding-top: 17px;
-      width: 80%;
-    }
-
-    .content {
-      @include media(tablet) {
-        width: 64.6%;
-      }
-
-      @media (max-width: 770px) {
-        width: 100%;
-      }
-
-      h2 {
-        a {
-          font-size: em(18, 16);
-          line-height: 1.11111111111;
-          padding: 3px 0 2px 0;
-
-          @media (max-width: 641px) {
-            font-size: 1em;
-          }
-
-          @include ie-lte(7) {
-            font-size: 0.75em;
-          }
-        }
-
-        @media (max-width: 641px) {
-          padding-bottom: 0;
-        }
-      }
-
-      p {
-        @include copy-16;
-        float: left;
-        min-width: 18em;
-      }
-
-      a {
-        text-decoration: none;
-
-        &:hover {
-          text-decoration: underline;
-        }
-      }
-    }
-
-    .organisations-box {
-      @include ie-lte(7) {
-        .organisations-box-inner {
-          padding-top: 0;
-        }
-      }
-
-      h3 {
-        @include copy-16;
-        border-bottom: none;
-
-        @media (max-width: 641px) {
-          padding-bottom: 0;
-        }
-      }
-
-      p {
-        float: none;
-        @include copy-14;
-      }
-
-      li {
-        list-style: none;
-        margin-bottom: 10px;
-        padding-bottom: 10px;
-        border-bottom: 1px solid $grey-2;
-
-        @include ie(6) {
-          margin-left: $gutter-half;
-          margin-right: $gutter-half;
-        }
-
-        &:last-child,
-        &.last-child {
-          margin-bottom: 0;
-          padding-bottom: 0;
-          border: none;
-        }
-
-        &.last-child {
-          @include ie(8) {
-            margin-bottom: 0;
-            padding-bottom: 0;
-            border: none;
-          }
-        }
-      }
-    }
-  }
-
-  .most-popular .heading-block .content h2 a {
-    color: $black;
-  }
-
-  .organisations-index .department .view-all {
-    margin-right: 0;
-    float: right;
-
-    &:after {
-      content: " \25B2";
-    }
-  }
-
-  .organisations-index .js-hide {
-    .organisations-box {
-      display: none;
-    }
-
-    .view-all:after {
-      content: " \25BC";
-    }
-  }
-
-  .document-list .document-row {
-    p {
-      float: none;
-    }
-
-    .contact-meta {
-      @include ig-core-14;
-      color: #6F777B;
-      display: block;
-      float: none;
-    }
-  }
-}
