--- app/assets/stylesheets/frontend/views/_documents-list.scss
@@ -1,11 +1,0 @@
-.document-list .document-row p {
-  font-family: "nta","GDS-Regular",sans-serif;
-  font-size: 14px;
-  line-height: 1.42857;
-  font-weight: 400;
-  text-transform: none;
-  float: left;
-  display: block;
-  list-style: none;
-  padding: 0 15px 0 0;
-}
