--- app/assets/stylesheets/frontend/views/_layouts.scss
@@ -1,73 +1,0 @@
-// Styles for Layouts view
-
-.clear {
-  clear: both;
-}
-
-@include media(desktop){
-  .block-1 {
-    width: 33.33%;
-    float: left;
-  }
-  .block-2 {
-    width: 66.66%;
-    float: right;
-  }
-}
-.block-pull-right {
-  float: right !important;
-}
-.block-pull-left {
-  float: left !important;
-}
-
-@include ie-lte(6){
-  .block-1 {
-    width: 32%;
-    float: left;
-  }
-  .block-2 {
-    width: 66.66%;
-    float: right;
-    overflow-x: hidden;
-  }
-}
-
-.inner-block {
-  padding: 0 $gutter-half;
-
-  @include media(tablet){
-    padding: 0 $gutter;
-  }
-
-  @include ie-lte(6){
-    padding: 0 $gutter;
-  }
-
-  &.floated-children {
-    @extend %contain-floats;
-
-    padding: 0;
-
-    @include media(tablet){
-      padding: 0 $gutter-half;
-    }
-  }
-}
-
-@include media(mobile){
-  .block-1 {
-    width: 100%;
-    float: none;
-    clear: both;
-  }
-  .block-2 {
-    width: 100%;
-    float: none;
-    clear: both;
-  }
-}
-
-.beta {
-  @include phase-tag(beta);
-}
