+++ app/controllers/admin_controller.rb
@@ -3,7 +3,6 @@ class AdminController < ApplicationController
 
   prepend_before_filter :authenticate_user!
   before_filter :require_signin_permission!
-  before_filter :skip_slimmer
 
   def info_for_paper_trail
     { user_name: current_user.name }
