--- app/views/contacts/_contact.html.erb
@@ -1,11 +1,0 @@
-<li class="organisation js-filter-item document-row" data-filter-terms="<%= contact %>">
-  <h3 id="<%= contact.slug %>">
-    <%= link_to contact, contact_path(organisation, contact) %>
-  </h3>
-  <p class="contact-meta">
-    <%= contact.contact_groups.join(", ") %>
-  </p>
-  <p>
-    <%= contact.description %>
-  </p>
-</li>
