--- app/views/contacts/_contacts_count.html.erb
@@ -1,6 +1,0 @@
-<span class="count">
-  <%= contacts.size %>
-</span>
-<strong>
-  <%= "Contact".pluralize(contacts.size) %>
-</strong>
