--- app/views/contacts/index.html.erb
@@ -1,73 +1,0 @@
-<% page_class 'contacts-page documents-index' %>
-
-<% content_for :extra_headers do %>
-  <meta name='description' content='Contact details and helplines for enquiries with HMRC' />
-<% end %>
-
-<header class="block headings-block">
-  <div class="inner-block floated-children">
-    <div class="heading-big">
-      <div class="inner-heading">
-        <h1>
-          <%= organisation.title %> Contacts
-        </h1>
-      </div>
-    </div>
-  </div>
-</header>
-
-<div class="block-1 filter-form">
-  <div class="inner-block">
-    <div class="filter-block">
-      <h2>
-        You can use the filters to show only results that match your interests
-      </h2>
-
-      <form method="get" action="<%= contacts_path %>"  class="js-filter-form" data-remote="true">
-        <fieldset>
-          <legend class="visuallyhidden">
-            Filter publications
-          </legend>
-          <div class="filter" id="keyword-filter">
-            <div class="filter-inner">
-              <%= label_tag "search[q]", "Contains", class: "title" %>
-              <%= text_field_tag "search[q]", search.q, type: "search" %>
-            </div>
-          </div>
-
-          <div class="filter">
-            <div class="filter-inner">
-              <%= label_tag "search[contact_group_id]", "Topic" %>
-              <%= select_tag "search[contact_group_id]", options_for_select(contact_groups.map {|c| [c, c.id]}, search.contact_group_id), { prompt: "All Topics", class: "single-row-select" } %>
-            </div>
-          </div>
-
-          <input class="button js-hidden" name="commit" type="submit" value="Refresh results">
-        </fieldset>
-      </form>
-
-    </div>
-  </div>
-</div>
-
-<div class="block-2 filtered-results index-list-page organisations-index">
-  <div class="inner-block">
-    <div class="filter-results-summary">
-      <span class="help-block hidden" id="working_on_filter_contacts">Loading results...</span>
-      <div class="selections contacts-count">
-        <%= render "contacts_count", contacts: @contacts %>
-      </div>
-    </div>
-    <div class="js-filter-block">
-      <div class="content">
-        <ol class="document-list">
-          <%= render @contacts %>
-        </ol>
-      </div>
-    </div>
-  </div>
-</div>
-
-<div class="js-filter-no-results">
-  <p>No contacts match that filter.</p>
-</div>
