--- app/views/contacts/index.js.erb
@@ -1,3 +1,0 @@
-$("#working_on_filter_contacts").addClass("hidden");
-$(".document-list").html("<%= j render @contacts %>");
-$(".contacts-count").removeClass("hidden").html("<%= j render 'contacts_count', contacts: @contacts %>");
