--- config/initializers/slimmer.rb
@@ -1,15 +1,0 @@
-Contacts::Application.configure do
-  config.slimmer.logger = Rails.logger
-
-  if Rails.env.production?
-    config.slimmer.use_cache = true
-  end
-
-  if Rails.env.development?
-    config.slimmer.asset_host = if ENV["STATIC_DEV"]
-      Plek.new.find('static')
-    else
-      "https://assets-origin.preview.alphagov.co.uk"
-    end
-  end
-end
